@@ -7,13 +7,18 @@ class Subject
 
   scope :root, -> { where(type: 'root').asc(:order) }
   scope :active_root, -> { where(type: 'root', status: 'active').asc(:order) }
+  scope :by_type, -> (type) { where(type: type) }
   scope :active_non_root, -> { where(:type.ne => 'root', :status => 'active') }
   scope :active, -> { where(status: 'active').asc(:order)  }
+  scope :not_bad, -> { where(:status.ne => 'bad').asc(:order)  }
+  scope :visible_marks, -> { where(:status.ne => 'bad', :region.ne => nil).asc(:order)  }
   scope :complete, -> { where(status: 'complete').asc(:order)  }
   scope :by_workflow, -> (workflow_id) { where(workflow_id: workflow_id)  }
-  scope :by_parent_subject_set, -> (parent_subject_set_id) { where(parent_subject_set_id: parent_subject_set_id)  }
+  scope :by_subject_set, -> (subject_set_id) { where(subject_set_id: subject_set_id).asc(:order)  }
   scope :by_parent_subject, -> (parent_subject_id) { where(parent_subject_id: parent_subject_id) }
+  scope :by_group, -> (group_id) { where(group_id: group_id) }
   scope :user_has_not_classified, -> (user_id) { where(:classifying_user_ids.ne => user_id)  }
+  scope :user_did_not_create, -> (user_id) { where(:creating_user_ids.ne => user_id)  }
 
   # This is a hash with one entry per deriv; `standard', 'thumbnail', etc
   field :location,                    type: Hash
@@ -21,7 +26,6 @@ class Subject
   field :status,                      type: String,  default: "active" #options: "active", "inactive", "bad", "retired", "complete", "contentious"
 
   field :meta_data,                   type: Hash
-  field :secondary_subject_count,     type: Integer, default: 0
   field :classification_count,        type: Integer, default: 0
   field :random_no,                   type: Float
   field :secondary_subject_count,     type: Integer, default: 0
@@ -34,10 +38,11 @@ class Subject
 
 
   # ROOT SUBJECT concerns:
-  field :order
+  field :order,                       type: Integer
   field :name,                        type: String
-  field :width
-  field :height
+  field :width,                       type: Integer
+  field :height,                      type: Integer
+  field :zooniverse_id
 
   # SECONDARY SUBJECT concerns:
   field :data,                        type: Hash
@@ -45,9 +50,11 @@ class Subject
 
   # Denormalized array of user ids that have classified this subject for quick filtering
   field :classifying_user_ids,        type: Array, default: []
-  field :deleting_user_ids,        type: Array, default: []
+  field :deleting_user_ids,           type: Array, default: []
+  field :creating_user_ids,           type: Array, default: []
 
   belongs_to :workflow
+  belongs_to :group
   belongs_to :parent_subject, :class_name => "Subject", :foreign_key => "parent_subject_id"
   belongs_to :subject_set, :class_name => "SubjectSet", :foreign_key => "subject_set_id"
 
@@ -61,12 +68,22 @@ class Subject
   after_create :update_subject_set_stats
   after_create :increment_parents_subject_count_by_one, :if => :parent_subject
 
+  # Index for typical query when fetching subjects for Transcribe/Verify:
+  index({"status" => 1, "workflow_id" => 1, "classifying_user_ids" => 1}, {background: true})
+  # Index for Marking by subject set:
+  index({"type" => 1, "subject_set_id" => 1}, {background: true})
+  # Index for fetching child subjects for a parent subject, optionally filtering by region NOT NULL
+  index({parent_subject_id: 1, status: 1, region: 1})
+  
+
   def thumbnail
     location['thumbnail'].nil? ? location['standard'] : location['thumbnail']
   end
-  
+
   def update_subject_set_stats
+    subject_set.subject_activated_on_workflow(workflow) if ! workflow.nil? && status == 'active'
     subject_set.inc_subject_count_for_workflow(workflow) if ! workflow.nil?
+    # subject_set.inc_active_secondary_subject 1 if type != 'root'
   end
 
   def increment_parents_subject_count_by(count)
@@ -128,16 +145,22 @@ class Subject
   def bad!
     status! 'bad'
     subject_set.subject_deactivated_on_workflow(workflow) if ! workflow.nil?
+    # subject_set.inc_complete_secondary_subject 1 if type != 'root'
   end
 
   def retire!
+    return if status == "bad"
+    return if classifying_user_ids.length < workflow.retire_limit
     status! 'retired'
     subject_set.subject_completed_on_workflow(workflow) if ! workflow.nil?
+    
+    # subject_set.inc_complete_secondary_subject 1 if type != 'root'
   end
 
   def activate!
     status! 'active'
-    self.subject_set.subject_activated_on_workflow(workflow) if ! workflow.nil?
+    subject_set.subject_activated_on_workflow(workflow) if ! workflow.nil?
+    # subject_set.inc_active_secondary_subject 1 if type != 'root'
   end
 
   def calculate_most_popular_parent_classification
@@ -155,6 +178,31 @@ class Subject
   def to_s
     "#{status != 'active' ? "[#{status.capitalize}] " : ''}#{workflow.nil? ? 'Final' : workflow.name.capitalize} Subject (#{type})"
   end
+
+
+  # Returns hash mapping distinct values for given field to matching count:
+  def self.group_by_field(field, match={})
+    agg = []
+    agg << {"$match" => match } if match
+    agg << {"$group" => { "_id" => "$#{field.to_s}", count: {"$sum" =>  1} }}
+    self.collection.aggregate(agg).inject({}) do |h, p|
+      h[p["_id"]] = p["count"]
+      h
+    end
+  end
+
+  # Same as above, but restricted to Group:
+  def self.group_by_field_for_group(group, field, match={})
+    self.collection.aggregate([
+      {"$match" => { "group_id" => group.id }.merge(match)}, 
+      {"$group" => { "_id" => "$#{field.to_s}", count: {"$sum" =>  1} }}
+
+    ]).inject({}) do |h, p|
+      h[p["_id"]] = p["count"]
+      h
+    end
+  end
+
 
   private
 
