@@ -5,7 +5,6 @@ module SubjectGenerationMethods
     def process_classification(classification)
 
       atts = subject_attributes_from_classification(classification)
-
       atts[:status] = 'inactive'
 
       classification.child_subject = Subject.find_or_initialize_by(workflow: atts[:workflow], parent_subject: atts[:parent_subject], type: atts[:type])
@@ -32,13 +31,44 @@ module SubjectGenerationMethods
       num_parent_classifications = classification.child_subject.parent_classifications.count
 
       # If subject has enough parent classifications, activate it:
-      puts "considering activating.... if #{num_parent_classifications} >= #{classification.workflow.generates_subjects_after}"
       if num_parent_classifications >= classification.workflow.generates_subjects_after
-        puts "Activating generated subject because now has #{num_parent_classifications} parent classifications"
-        atts[:status] = 'active'
+
+        # Get number of distinct classifications:
+        num_vals = classification.child_subject.data['values'].nil? ? -1 : classification.child_subject.data['values'].size
+
+        # Where will this generated subject appear, if anywhere?
+        next_workflow = classification.child_subject.workflow
+
+        # If there is no next workflow, this subject is done. Presumably the retire_limit caused the parent subject to be retired as well.
+        if next_workflow.nil? 
+          atts[:status] = 'complete'
+
+        # There is a next workfllow (probably Verify)
+        else
+          # Get subject-generation method type (presumably for Verify workflow) (which is likely 'most-popular')
+          verify_method = next_workflow.generates_subjects_method
+
+          # If next workflow's generation method is most-popular and everyone transcribed the same thing, auto upgrade to 'complete':
+          if num_vals == 1 && verify_method == 'most-popular'
+            atts[:status] = 'complete'
+
+          # .. Otherwise, activate the generated subject into the next workflow:
+          else
+            classification.child_subject.activate!
+            atts.delete :status
+          end
+        end
       end
 
-      puts "Saving atts to classification: #{atts.inspect}"
+      # PB: At writing, only verify uses collect-unique. It's important that
+      # subjects generated from transcribe not be classifyable (i.e. voted upon)
+      # by any user submitting a transcription. We should probably support a 
+      # thus generated are not classifyable by classification authors.
+      atts[:creating_user_ids] = classification.child_subject.creating_user_ids
+      atts[:creating_user_ids] ||= []
+      classification.child_subject.creating_user_ids.push classification.user_id
+
+      # puts "Saving atts to classification: #{atts.inspect}"
       classification.child_subject.update_attributes atts
 
       classification.child_subject
